backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flagembedding>=1.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "elevenlabs>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]
